refactor: emulator 관련 엔티티를 database 모듈로 이동
