+++ database/src/main/java/com/pickcar/emulator/domain/GpsStatus.java
@@ -1,4 +1,4 @@
-package com.pickcar.domain;
+package com.pickcar.emulator.domain;
 
 public enum GpsStatus {
     NORMAL, ABNORMAL, NOT_INSTALLED, ON_ABNORMAL
