+++ domain/src/main/java/com/pickcar/drivehistory/application/DriveHistoryService.java
@@ -4,8 +4,8 @@ import com.pickcar.application.CycleService;
 import com.pickcar.application.EventInfoService;
 import com.pickcar.auth.application.UserService;
 import com.pickcar.auth.domain.User;
-import com.pickcar.domain.Cycle;
-import com.pickcar.domain.EventInfo;
+import com.pickcar.emulator.domain.Cycle;
+import com.pickcar.emulator.domain.EventInfo;
 import com.pickcar.drivehistory.domain.DriveHistory;
 import com.pickcar.drivehistory.infrastructure.DriveHistoryRepository;
 import com.pickcar.reservation.application.ReservationService;
